Extract hetzner.py module_utils.
